% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalability.R
\name{scalability}
\alias{scalability}
\title{Mutate a dataframe to add a scalability column}
\usage{
scalability(df)
}
\arguments{
\item{df}{The dataframe that contains time and processor count information}
}
\value{
A new dataframe with scalability
}
\description{
Mutate a dataframe to add a scalability column
}
\examples{
df <- data.frame(p=c(1,2,4))
df$t <- runif(nrow(df), 1.0, 2.0) / df$p
scalability(df)
}

